Remove check for C_UNDERSCORE_SYMBOLS, which isn't defined anymore
In the big build system rewrite, C_UNDERSCORE_SYMBOLS was removed. The
reason was that the check for it was actually not portable or reliable.
At the same time, no currently supported system seems to require us to
prepend symbols we reference in loadable modules with an additional
underscore prefix.

If we ever get reports about a system not working due to this, we can
add support for this back, but then with a clear documentation as to
which system actually needs it (note that the removed code lists several
operating systems that do NOT need the underscore, but none that do.