Simplify OBJ_REFLVAR
This actually reverts the code to how it was before 1997-03-06, when
it was changed as part of the merge of compiler branch. My guess is that
this was a kind of micro optimization.

However, the loss in clarity it needs IMHO makes it worthless. It also
will break down if we ever try to change the internal layout of these
data structures, as it has implicit dependencies on the exact way
OBJ_LVAR and LVAR_REFLVAR resp. REFLVAR_LVAR are implemented.