Tweak FuncINTLIST_STRING code
Using a 1K buffer just to cache the result of INTOBJ_INT (which is
trivial to compute using a shift and setting a bit) hurts more than it
helps on modern CPUs (looking at the machine code, we replace a memory
lookup by an addition, as a shift has to be done in either case).

So we instead remove the cache, and code dealing with it.