Fix regression in DoImmutableMatrix, add tests
A previous commit added a check for IsFFECollection(field) (which
was suggested by me), but that is not actually correct, as "field"
can also be an integer. Instead, check IsFFECollColl(matrix).

Also fix a typo, change tabs to spaces, and refactor the code a
little bit.

And finally, add more systematic tests for ImmutableMatrix