Fix SySetGapRootPath when multiple paths separated by ; are given
This fixes a regression introduced when I integrated HPC-GAP: Back then,
I changed SySetGapRootPath to add two root paths for each path
given to it: Given path XYZ, it adds `XYZ/hpcgap/` and then `XYZ`. This
allows us to transparently override parts of the library for HPC-GAP.

However, I neglected to test with arguments to the `-l` command line
option which contain a semicolon to separate multiple paths, and of
course I broke that. This commit fixes this regression.