traverse.c: replace TraversalMask by a method enum
The idea of using a mask for traversal is cute, but had the unfortunate
drawback that for objects using TRAVERSE_ALL or TRAVERSE_ALL_BUT, in fact
not all sub-objects were traverse, but rather at most 31 resp. 63 of them
We could have fixed that by negating the meaning of bits (so that a zero
bit indicates "traverse this", relying on zero bits being shifted in.

Alas, instead we simply switch to a fixed list of four traversal methods.
The resulting code is simpler, and virtually anything can be done via
TRAVERSE_BY_FUNCTION anyway.