Remove GASMAN's 'dirty mode'
NewBag now always returned zeroed bags, and promises so in its
documentation. Rationale: Quite some code implicitly relies on this
behaviour by now. So if we wanted to change it, we'd have to do more
than just flip on "dirty mode" (resp. remove the relevant memset in the
GASMAN collector code). We'd then have to review all code making
allocations as to whether it relies on the zeroing behaviour or not (and
that can be quite tricky to refute). The existing "dirty mode" would not
help with that in any serious way, so we may just as well remove it now.