Avoid floating point computations in Test()
... for computing progress percentage. Note that the new code
intentionally always rounds down: We don't want to show 100%
if we are not on the last line. (And that last line is usually
doing STOP_TEST(), so you may never see 100%, but that's fine).

This fixes a bad interaction with the tests of the float package,
which temporarily switches GAP to different float handlers, some
of which may not implement Round().