Prevent Z(p,d), p non-prime, from breaking the type system
Invoking Z(p,d) with p not a prime used to crash gap, which we fixed.
However, invocations like `Z(4,5)` still would erroneously trigger the
creation of a type object for fields of size p^d (in the example: 1024),
with the non-prime value p set as characteristic. This could then corrupt
subsequent computations.

Also get rid of a stray #include