kernel: turn CountGVars into an T_INT object
This allows us to save it in workspaces, which makes it possible to
enable USE_GVAR_BUCKETS and still use workspaces.

The drawback of this is that we have to insert INT_INTOBJ calls in a few
places, but these amount to a bitshift, which is cheap. The only
possibly somewhat critical code path using CountGVars is GVarName(). And
there, we put the shifted value into a temporary local variable. This
enables substantial additional compiler optimizations (previously, the
compiler had to frequently reload CountGVars, as it could not deduce
that no other function or thread will modify it). In total, this means
things get very slightly faster with this commit, not slower.