Fix some referencing issues in HPC-GAP documentation
References to CallWithTimeout and CallWithTimeoutList have been
converted to text, since these functions no longer exist.

Sometimes multiple sections describe functions with the same name,
leading to collisions; we change the labels on some so labels aren't
multiply defined.

A reference to the section on write-once functions was bungled, this has
been corrected.