intobj: add comments to prod_int, avoid arithmetic right shift
The old code for checking whether an integer fits into a half a word
implicitly relied on right shifts being arithmetic (i.e., replicating
the sign bit).

With GCC on x86, old and new code produce the same machine code, but the
new one is (a) correct, and (b) much easier to understand.