Add BREAKPOINT(n) function.
The purpose behind this function is to make it easier to set debugger
breakpoints in the kernel dependent on the state of the GAP code.

One use case is to simply insert `BREAKPOINT(0)` at a specific GAP
source location and then set a breakpoint on `FuncBREAKPOINT` in the
kernel.

The function accepts any argument; if it is a small integer, it will be
converted and stored in the global variable `BreakPointValue`. This
also allows the encoding of GAP state in that value and to attach a
condition based on the value of `BreakPointValue` to other breakpoints.