Reimplement WITH_IMPS_FLAGS and CLEAR_IMP_CACHE in C
This commit is a first step towards getting rid of c_filter1.c.
It implements the functions `WITH_IMPS_FLAGS` and `CLEAR_IMP_CACHE`
from `lib/filter1.g` as kernel functions.
The data for the cache and, with COUNT_OPERS option, the counter
variables are now in the kernel.
They can be imported to GAP level with `WITH_IMPS_FLAGS_STAT` and
`OPERS_CACHE_INFO`.
Missing counters for `WITH_HIDDEN_IMPS_FLAGS` were also added.
The list `IMPLICATIONS` is now generated in the kernel.

In `WITH_IMPS_FLAGS` we avoid to run through the whole list of
IMPLICATIONS in the last round.

On my machine I get these runtimes for
   TestDirectory( [ DirectoriesLibrary( "tst/testinstall" ) ]);

With lib/filter1.g (no src/c_filter1.c)     50937 msec
With src/c_filter1.c                        46304 msec
With this patch                             43860 msec