Store declaration filters of operations in record
This substitutes a list OPERATIONS containing operations in odd positions
and corresponding lists of flag lists from declarations in the following
positions by:
- a list OPERATIONS, containing only the operations themselves and
- a record OPER_FLAGS containing the lists of flag lists and infos about
  the location of declarations.

The locations may be used with `PageSource` and other debugging tools.

More precisely, a tuple [oper,flaglists,locations] is stored in component
MASTER_POINTER_NUMBER(oper) of OPER_FLAGS.
We need a backpointer to oper here because MASTER_POINTER_NUMBER(oper) will
change after reloading a workspace. With the backpointer we can
clean this up in a post-restore function.

On my computer the startup time of GAP is reduced from about 1460msec
to 1260msec.