Load PrimGrp, SmallGrp and TransGrp packages if available
If any of these packages is not available, GAP will read corresponding
directory (`prim`, `small` or `trans`).

Also replaced the call of `IsIrreducible` in `lib/galois.g` by
`IsIrreducibleRingElement` to allow reading it earlier than
`overload.g`, and moved definition of `TRANSREGION` out of `trans`
directory to make it working both with `trans` and TransGrp package.