fixed a bug in `RadicalOfAlgebra`
For associative algebras that are not matrix algebras,
the radical is computed in an isomorphic matrix algebra.
Up to now, one ran into an error if `Coefficients`
returned an immutable list.

(One might argue that `Coefficients` should always return
a new mutable list, but this would cause the creation of
many superfluous lists.
Thus we solve the problem in the place where the list
has to be changed.)