PageSource for kernel functions and operations
PageSource now tries to show the source of kernel functions.

It will show the correct file (if available), and try to find a line
containing

Obj FuncXYZFUNCTION

for the function `XYZFUNCTION`.
(The builtin pager just shows the file.)

This will currently not work for kernel functions in packages because
their FilenameFunc ist just a file name, without directory relative to GAP
root.

For operations the function now shows the first place where it was declared
and writes information about the number of declarations. `PageSource` can be
used with a second argument to get to the other locations of declarations.