Just use the generic list access code in ASS2_LIST
The original code lead to the following error

```
gap> l := [[]];
[ [  ] ]
gap> l[1,2] := 4;
4
gap> l;
[ [  ] ]
```

This is caused by the list l[1] being empty, and the optimised code
in ASS2_LIST calling AssPlist on this list, not retyping it correctly.

Just deleting the optimisation the code behave safely (albeit maybe
slightly slower).

An alternative fix would be to add a check whether the list is
empty and call the appropriate assignment functions, or edit
AssPlist to correctly assign to the empty list.