Use generic list access code in ASS2_LIST
The original code lead to the following incorrect behavior

```
gap> l := [[]];
[ [  ] ]
gap> l[1,2] := 4;
4
gap> l;
[ [  ] ]
```

This is caused by the list l[1] being empty, and the optimised code
in ASS2_LIST calling AssPlist on this list, not retyping it correctly.

A related problem occurs when l[1] is immutable: the code ignored that,
and would modify l[1] anyway.

Just deleting the optimisation the code behave safely (albeit is
slightly slower).