grp.gi: add special case to MinimalGeneratingSet
Currently, `MinimalGeneratingSet` only supports soluble groups. Cyclic
groups are soluble, and so a group that is not soluable requires at
least two generators.  Therefore, in `MinimalGeneratingSet`, we can add
a special case for groups that have two generators.  Once we have found
that they are not soluble, if they only have two generators, then that
generating set must have minimal size, and we can return it.