semigrp/tran: fix diffs when Semigroups is loaded
In the test files `tst/testinstall/semigrp.tst` and `semitran.tst`,
there are some diffs when the Semigroups package is loaded.
In these tests, there are semigroups that happen to be groups. When
the Semigroups package is loaded, there are additional methods that
realise this information, and so the ViewString is different,
causing the diffs.

I have changed an ImmediateMethod to `lib/semigrp.gd` so that these
semigroups know that they are groups, even when the Semigroups package
is not loaded.

In the future, we will be running the `GAP` test files in the continuous
integration for the Semigroups package, and so diffs like this should be
caught immediately.

There still remains one diff in `tst/testinstall/semigrp.tst` when the
Semigroups package is loaded. This is an actual bug, and is being
sorted in the Semigroups repo.