Add GetWithDefault
GetWithDefault provides a way of checking if an index in a list
is bound, and the value if the index is bound, as one function call.

It is particularly useful in weak pointer lists and atomic lists,
where there is a race condition between checking if an element is
bound, and then reading it's value.