Introduce GAP-Level StandardInput, StandardOutput, StandardError
This introduces the InputStream StandardInput and the OutputStreams
StandardOutput and StandardError at the GAP level, and attaches
 * InputTextFile("*stdin*") to StandardInput,
 * OutputTextFile("*stdout*", true) to StandardOutput
 * OutputTextFile("*errout*", true) to StandardError

This small change already enables refactoring error.g to PrintTo
the StandardError stream, and it can be replaced by an
OutputTextString to capture error messages.

One obvious use case of this is to capture output for alternative
REPL implementations such as Jupyter.