Improve performance of Info, and merge with HPC-GAP
This patch moves IsInfoClassRep information out of a global
structure into the class, to improve performance.

Also add a special case at the Kernel level to optimise the most
common situation.

This reduces the speed gap between Info and an equivalent if
statement from 10x to ~1.5x

Also remove the need for a seperate HPC-GAP implementation