semirel: add family check to GreensXClassOfElement
This commit adds a check that the family of the object and the elements
family of the family of the semigroup passed to the methods
GreensXClassOfElement, where X in [R, L, H, D, J], are equal. This
prevents the creation of Green's classes with wholly the wrong type of
objects as representatives.