hpcgap: tweak MASTER_POINTER_NUMBER
In HPC-GAP, we do not have actual master pointer numbers, and need to
fake something. So far, we used HANDLE_OBJ, but that can return values
that don't fit into an immediate integer. To reduce the chance for that
a bit, we now divide the result by sizeof(Obj), which is safe, as all
objects are aligned to multiples of that value, i.e. the lower bits are
unused.