semigrp: fix DisplaySemigroup
The documentation for `DisplaySemigroup` was not linked.
Furthermore, there was an unexpected error for certain
types of transformation semigroup when using `DisplaySemigroup`:
in particular, the method did not take into account the degree
of the semigroup when trying to calculate the rank of an
element. For example, in any monoid of transformations of degree `n`,
`IdentityTransformation` has rank `n`. So you need to know the
degree of the semigroup to get the rank.