Modify READ_ALL_COMMANDS to copy results
this is done to cater for the following situation: In Jupyter a user
wants to execute the following code in a single cell:
```
a:=[1,"a"];
b:=a;
b[2]:=3;
a;
```

Before this patch, the result is

```
[ [ true, [ 1, 3 ] ], [ true, [ 1, 3 ] ], [ true, 3 ], [ true, [ 1, 3 ] ] ]
```

after it is

```
[ [ true, [ 1, "a" ] ], [ true, [ 1, "a" ] ], [ true, 3 ], [ true, [ 1, 3 ] ] ]
```

and yields the output the user expects.

Note that copying the results of expressions is potentially expensive, but this
problem needs to be addressed by making the scanner/parser/interpreter capable
of partially reading an input stream and not forgetting its (remaining) contents.