Avoid segfault if custom function call method fails to return value.
This is a follow-up to the change moving the SET_BRK_CALL_TO() invocations, to
also share the postprocessing after DispatchFuncCall with the "traditional"
function call. This has the benefit of preventing e.g. segfaults should the
DispatchFuncCall fail to return an object when it is supposed to.

In addition, added tests to tst/testinstall/callfunc.tst to cover the behavior
of GAP in such error cases as just mentioned. (Thanks @fingolfin for the
guidance on the structure of these tests.)