Removed provisional code for coexisting data libs and packages
PR #1714 added some code needed to load either a new package for
a data library, if present, or load the existing prim, small or
trans directory otherwise. This code is not needed any more,
since we remove prim, trans, and small directories.

Also make PrimGrp, SmallGrp and TransGrp packages needed to run
GAP. This ensures that a proper error message will be displayed
in case any of these packages is missing.