Make line editing resilient to LineEditKeyHandler failure
Specifically, if the LineEditKeyHandler returns nothing (e.g. due to an
Error()), or if it returns something that isn't a list, or a list of the wrong
size, then just ignore this, instead of running into segfaults.

This addresses one part of issue #1853

Also remove superfluous "else {" and "}" when HAVE_LIBREADLINE is set.