Simplify caching logic in GeneralMappingsFamily
Previously, the cache consisted of a weak list with entries alternating
between "keys" (= range domain) and "values" (= the corresponding mapping
family).

But the "values" contain the keys (via FamilyRange(fam). By exploiting that,
we can simplify the caching logic. The resulting code is very similar to the
code in DirectProductElementsFamily, suggesting that we can eventually replace
both with a common abstraction.