Omit non-characters from `PermChars` results
Up to now, it might happen that the result of `PermCandidatesFaithful`
contains virtual characters that are not characters,
provided that one prescribes a list of rational irreducible constituents
and that possible permutation characters exist
which have negative scalar products with some rational irreducible
characters that are not among the given constituents.

These cases appear to be rare,
and the smallest example which I know runs about 20 minutes.

I regard the change as a bugfix, since the documentation states
that the result consists of characters.