HACK: extend (Test)BindOnce(Expr) to plists
This works since plist and positional objects have identical
memory layout. One caveat: the TNUM is now adjusted, so if it
was T_PLIST_EMPTY, it stays that way. This should be fixed
(basically, if the input TNUM is neither T_POSOBJ nor T_PLIST,
set it to T_PLIST).

Alternatively, we turn this into an internal-only function;
because for our us in MemoizePosInt, the TNUM does not matter.