Use findValueInNams in ReadCallVarAss
Note that before with change, we searched through 'nams' in descending order,
and now we search in ascending order. But since 'nams' cannot contain any
duplicate entries, this is irrelevant.

Remove code which limited nest0 to < 2^16 -- the actual limitation was removed
in a commit dated 2005-10-17. Now it is limited by sizeof(UInt), which for our
purpose is the same as being unlimited.

Also remove the limitation on indx to be less than 1024. This limited debug
variables, where we only considered at most 1024 entries in 'nams'. This was
an artifact from a time when this was the actual limit. But the true limit
since the same commit dated 2005-10-17 mentioned above, is 2^16 = 65536.
And we already enforce that limit when parsing functions, plus a paranoia
check for it in findValueInNams().