Speed up Factorial
Commit 7a1953c made BINOMIAL_INT from the gmp library available.

I have GAP code that is almost as fast as BINOMIAL_INT for many integers and
even faster for big arguments. This was never commited because the code
produces and caches long lists of primes.

Nevertheless, with any fast implementation of Binomial one also gets a
faster implementation of Factorial. This is provided by this commit.