Move INFODATA_* constants into the kernel
We previously hard coded the assumption that INFODATA_CURRENTLEVEL is 1
right into the kernel, without any hint about this, so a change would
have broken it.

By defining INFODATA_CURRENTLEVEL in an enum in the kernel, this cannot
happen anymore.

Also used that opportunity to actually reorder the INFODATA_* values,
and put INFODATA_NUM first.