Get rid of SyStackAlign and SYS_STACK_ALIGN
Now there is only `C_STACK_ALIGN` and `StackAlignBags` left. The former is a
constant computed by autoconf and placed into config.h. This value is then
passed to `InitBags`, which stores it into StackAlignBags.

I decided to leave these two separate, as the `InitBags` documentation
provides lots of details on how this value is used, and I felt it easiest to
leave that as it is, instead of trying to rewrite it to be correct if I had
removed the `stack_align` argument to `InitBags`.

Also, my main goal is achieved: Now it only takes one step find out where the
value in StackAlignBags comes from, instead of three (or two instead of four,
depending on how you count). Much better, in any case.

Resolves #1218