Change all StructInitInfo instances to C99 init
Doing this here and also in packages opens up the possibility for us to
remove and add entries in the future (e.g. to remove revision_c, revision_h,
version, and instead add e.g. initModuleState).

It is also IMHO easier to read and get right, thus avoids errors. Indeed,
people already used comments to make it clear which field each line was
corresponding to, we just formalize this.