kernel: replace UNGET_CHAR by PEEK_CHAR
`UNGET_CHAR` was easy to use incorrectly. Indeed, I claim that three of the
four use cases were formally incorrect. Specifically, we did
`UNGET_CHAR(*STATE(In))`, but that puts the currently visible character back
-- it should put the one before that back! In one case, this did not matter,
as both of those were equal to '.'. This leaves two cases in which I think we
put back the wrong character, both are inside the floating point code. It
would be interesting to figure out if there is an input where this leads to
an incorrect reaction by GAP (I only briefly tried to find some, but all I
constructed ended up being inputs which leads to errors either way).

To fix this and avoid future problems, I replaced `UNGET_CHAR` by a new
function `PEEK_CHAR`. This has the added advantage that we know we'll never
need to put back things twice in a row. I.e. you could `PEEK_CHAR` multiple
times in a row without problems.