FIX/ENHANCE: Performance and memory for ImageKernelBlockHomomorphism.
The old code added every generator found with
`AddGeneratorsExtendSchreierTree'. This caused the storage of lots of
generators and subsequent memory issues.

Also, if there are few large blocks, Butler's block homomorphism code is
inefficient and it needs to run through all points in one block. In this
case an ordinary orbit calculation is far more efficient.

Together these issues did cause severe problems in large degrees with few
blocks (observed by Thomas).

It it possible that it also rectifies some of the observed problems in
larger degree.

What one could still do is to use random subproducts instead of testing all
Schreier generators, but this is not the right time to do so.