Allow Int(""), reject null bytes, clarify docs
Strings which contain a null byte are now rejected by `Int`,
instead of ignoring them and any characters after them.

Moreover, `Int("")` and `Int("-")` again returns 0 (as was the case in
all previous GAP versions). I allowed both because I realized that also
`Rat` accepts empty strings (i.e. `Rat("/1")` and `Rat("-/1")` both
return `0`, and this way, we can write a concise uniform definition: we
expect an optional sign character, followed by a sequence of decimal
digits -- and that sequence can of course also be empty.

In the opposite direction, we now forbid `Int("--")` and `Int("--1")`.

Also clarify the documentation of `Int`, `Rat` and `IntHexString` for
strings: The primary reference entry for `Int` now states what happens
in case of invalid input. It also now links to the secondary entry "Int
for strings".

The documentation for `Int`, `Rat` and `IntHexString` for strings was
previously jumbled together. Split this into three separate entries, and
be more precise in documenting which inputs are accepted, and what
happens for invalid inputs.

Fixes #2047