ENHANCE: When computing an isom. to a new fp group from a finite group
with `IsomorphismFpGroup`, store the *inverse* of this isomorphism as a
known faithful map to a better behaved group, thus avoiding such a searched
renewed if any "finite group" calculation is done in the FpGroup (not that
this is advisable, but users do it).

Ditto translate a known faithful map after calling
`(Isomorphism)SimplifiedFpGroup`.

This is suggested by #2020

also bail out in complicated case instead of trying hard

The output of a `Random` call in test file has changed.

REBASE: Comment cleanup