Fix SIZE_OBJ kernel func for large or immediate objects
That is, use ObjInt_UInt instead of INTOBJ_INT to convert the object size
into a GAP integer object, and handle T_INT and T_FFE inputs explicitly.

Also replace the kernel function SHALLOW_SIZE with a synonym pointing to
SIZE_OBJ. While the former name may be less misleading (a user might expect
SIZE_OBJ to include the sizes of subobjects), the latter name matches the
name of the kernel function, just like with TNUM_OBJ and TNAM_OBJ, which
simplifies life for kernel developers who want to use these UNDOCUMENTED,
INTERNAL and potentially DANGEROUS APIs. As such, clarity of names is not a
primary concern.