Rename OPER_FLAGS to OPER_DATA and STORE_OPER_FLAGS to STORE_OPER_DATA
The name is slighly more fitting (because the declaration locations are
stored in OPER_DATA), and one could store more information in OPER_DATA,
such as the type of the operation (Attribute, Property, Tester, Setter,
mutability of an attribute).

Note that GET_OPER_FLAGS is not renamed, because it still just returns
the flags.