lib: use IS_AND_FILTER to detect AND filters
This is slightly more efficient than searching through the FILTERS list,
and also requires no locking.

Note that on HPC-GAP, this also fixes an access violation bug, as we
should have taken a lock to read FILTERS. However, due the ward bug,
this silently worked, because ward failed to emit the required guards!