gac: suppress warnings about undefined vars
Previously, if one took a GAP function with an unused local, then in the C
code generated by gac for that function, there would be a corresponding
unused local variable, leading to a C compiler warning (turned into an
error in our CI tests). While that might be marginally useful (for the
same reasons an "unused local variable" warning can be useful in C code),
it can be problematic for code where a local *is* used, but only
conditionally, depending on a global constant (so e.g. inside of an `if
IsHPCGAP then ... fi;` statement).

We now emit some additional C code for each local to prevent the compiler
from warning about it being unused.