add shortname to IsomorphismTypeInfoFiniteSimpleGroup value
The record returned by `IsomorphismTypeInfoFiniteSimpleGroup`
contains a `name` component that is not suitable for
programmatic access.

The new component `shortname` fits to the names that are used in
`CTblLib` and `AtlasRep`.
It will simplify some of my example sessions.
In particular I will be able to get rid of some calls to
`StructureDescription`; I think it is not appropriate to advertise
that function by applying it to simple groups.

    TB