kernel: OnSetsPerm needs sets as input; don't remove dups
Note that we also didn't remove duplicates in the case were we acted on a set
of small integers, this duplicate removal step only happened when acting on
larger integers, permutations etc. So now we just never remove them; and as
long as the input is a set, we are fine. Luckily, `FuncOnSets` ensures the
input really always is a set.