kernel: overhaul OnSets & OnTuples for perms, pperms, trans
The code for these functions used LEN_LIST but elsewhere was written in such
a way that it was only corrected it the input list were non-empty plain (and
possibly having other properties). Luckily, the only callers for the six
involved functions already enforced that.

Documented this requirement, and then changed the code to explicitly use
plist accessor functions. Also use SortPlistByRawObj resp. SortDensePlist
where it (now) makes sense.

Also change FuncOnPosIntSetsPartialPerm and FuncOnPosIntSetsTrans in a
similar fashion.