Tweak tst/testinstall/coll.tst to work with semigroups pkg
An immediate method for IsTrivial provided by semigroups causes Magma(1) to
learn prematurely that it is trivial. Avoid this by using a magma (or, in the
end, monoid) generated by a singular matrix. Also, use Magma(0) instead of
Magma(1) in one other test.

On the long run, we might want to replace all uses of Magma() in this test
by custom ("mock") collection objects, as there is always a chance that a
(perfectly valid) new method in some package throws some of them off. But for
now, just using Magma() to create our tests objects is very convenient.