kernel: ensure uniform printing of macfloats nan, inf, -inf
On some platforms, NaN values might be printed as `nan`, on others as `NaN`;
on some, the sign of the Nan value may be printed, on other not so (and the
sign of a NaN is something which highly depends on the float implementation to
start with).

This made it difficult to write tests which rely on the way these values are
printed. Hence we now deal with these values explicitly and in a manner which
produces identical results across all platforms.

Fixes #2193