kernel: rewrite rest of method selection code in C
This is mostly about simplifying the code, and reducing code
duplication, and less about performance (there seems to be a negligible
difference with that).

In this first commit, the new method lookup code is added, but does not
replace the old code; instead, the results of the new and old code are
compared. This successfully passes the testinstall test suite, giving
confidence that the new code is correct.