Move unused code from grp/classic.gi to dev/deleted-code/
This code is never called by anything, nor is it documented. For some
of it (matrix group wreath products), newer code exists in the library.

Since the code may become relevant again for Alexander Hulpke one day,
let's not remove it right away, but instead move it to dev/deleted-code/

Resolves #2327