doc/tut: turn example using 'time' into a <Log>
This way, it is not run as part of the tests, which could cause problems
in the rare cases where the second `time` return 1 instead of 0, forcing
one to restart the affected CI job.

Additionally, I think it's more helpful for the reader of the example to
see an actual time for both, so they get the an idea about the kind of
speed up to expect (previously, they just saw that it changes from "at
least one millisecond" to "less than a millisecond").