kernel: remove precords code in ariths.c
The functions InstallOneObject, InstallSumObject, etc. are meant to
switch between normal and verbose operations for arithmetic operations
on external objects (i.e. positional and component objects, as well as
datobjs, weakpointer objects and possibly those provided by kernel
extensions in packages).

But for some reason, they also did that for precords. My guess is that
this was for backwards compatibility for GAP3 code, but AFAIK we removed
most of that elsewhere, so this code in ariths.c doesn't seem to serve a
purpose anymore; hence we remove it.

Note that even if my understanding is flawed, the impact of this change
should be harmless: It only affects tracing of such methods, but not
their actual use.