kernel: remove S_PARTIALFLOAT1, mark float parsing hack
Parsing float literals that start with a dot, like `.123`, is
implemented using a hack, where the reader tells the scanner that it
should look for a number expression (normally, the scanner never changes
the state of the reader, except for parsing long integers and float
literals).

We change the way this is done to make it stand out more. Further
benefits from the change are that we can get rid of S_PARTIALFLOAT1,
simplifying the intricate state machine for parsing. Moreover,
STATE(Symbol) now is read-only in the reader.